mudanças de organização de código
