--- src/carrinho.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import './App.css';
-// import styled from 'styled-components';
-
-// const carrinhoTitulo = styled.div`
-// border: 10px solid black;
-// `
-
-
-export default function Carrinho() {
-	return
-}
-
-
-{/* <button alt={''} src={props.icone} onClick={props.onClickIcone}/>
-		<p>{props.valorContador}</p> */}
