+++ src/components/Carrinho/carrinho-valor-total.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import '../../src/../App.css';
+
+export default function ValorTotal() {
+    const valor=200
+	return <div>
+		<span>
+            <p>Valor total:R${valor},00</p>
+        </span>
+        
+        
+	</div>
+}
